requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
html5lib
beautifulsoup4
ruamel.yaml
pydantic
cowsayg